% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build}
\alias{build}
\title{Title}
\usage{
build(featIDs, meta.data, dataAR, obffa.res, synex.res)
}
\arguments{
\item{featIDs}{Text}

\item{meta.data}{Text}

\item{dataAR}{Text}

\item{obffa.res}{Text}

\item{synex.res}{Text}
}
\value{
Text
}
\description{
Title
}
