% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exIPA.R
\name{exIPA}
\alias{exIPA}
\title{Title}
\usage{
exIPA(final.labels, col.ft.name, obffa.res, FtD.method)
}
\arguments{
\item{final.labels}{Text}

\item{col.ft.name}{Text}

\item{obffa.res}{Text}

\item{FtD.method}{Text}
}
\value{
Text
}
\description{
Title
}
