% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fExpr.R
\name{fExpr}
\alias{fExpr}
\title{Feature Expression extraction}
\usage{
fExpr(Data, colnum1, colnum2, n.ctrl, n.facA, n.facB, n.facAB)
}
\arguments{
\item{Data}{Text}

\item{colnum1}{Text}

\item{colnum2}{Text}

\item{n.ctrl}{Text}

\item{n.facA}{Text}

\item{n.facB}{Text}

\item{n.facAB}{Text}
}
\value{
Text
}
\description{
Text
}
