% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FtDbh.R
\name{FtDbh}
\alias{FtDbh}
\title{Feature Discovery with BH-method by OBIF}
\usage{
FtDbh(final.results, l2fc.co, pval.co)
}
\arguments{
\item{final.results}{Text}

\item{l2fc.co}{Text}

\item{pval.co}{Text}
}
\value{
Text
}
\description{
Feature Discovery with BH-method by OBIF
}
