% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FtDpv.R
\name{FtDpv}
\alias{FtDpv}
\title{Feature Discovery with p-value by OBIF}
\usage{
FtDpv(final.results, l2fc.co, pval.co)
}
\arguments{
\item{final.results}{Test}

\item{l2fc.co}{Text}

\item{pval.co}{Text}
}
\value{
Text
}
\description{
Feature Discovery with p-value by OBIF
}
