% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaD.R
\name{metaD}
\alias{metaD}
\title{Meta data extraction function}
\usage{
metaD(Data, colnum1, colnum2)
}
\arguments{
\item{Data}{Text}

\item{colnum1}{Text}

\item{colnum2}{Textdocu}
}
\value{
Text
}
\description{
Meta data extraction function
}
