% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normD.R
\name{normD}
\alias{normD}
\title{Normalization for Data matrix m}
\usage{
normD(expr.val, QCplot, col.samples)
}
\arguments{
\item{expr.val}{Text}

\item{QCplot}{Text}

\item{col.samples}{Text}
}
\value{
Text
}
\description{
Normalization for Data matrix m
}
