% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nTool.R
\name{nTool}
\alias{nTool}
\title{Sample number-based parameter making function for OBIF}
\usage{
nTool(param, n.ctrl, n.facA, n.facB, n.facAB)
}
\arguments{
\item{param}{Text}

\item{n.ctrl}{Text}

\item{n.facA}{Text}

\item{n.facB}{Text}

\item{n.facAB}{Text}
}
\value{
Text
}
\description{
Text
}
