% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obffa.R
\name{obffa}
\alias{obffa}
\title{Full factorial analysis}
\usage{
obffa(dataAR, design, contrast)
}
\arguments{
\item{dataAR}{A numeric data matrix with the analyis-ready feature expression values where row are features and columns are samples}

\item{design}{A design matrix to be applied for expression analysis}

\item{contrast}{A contrast matrix to be applied for contrast analysis}
}
\value{
A data frame with the individual p-values per feature for main effects, simple main effects and interaction effects.
}
\description{
The obffa function is the primary function of OBIF and is used to perform full factorial analysis in an Ananlysis-Ready dataset.
}
