% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obint.R
\name{obint}
\alias{obint}
\title{Omics-based interaction analysis by OBIF}
\usage{
obint(dataAR, design2x2, result)
}
\arguments{
\item{dataAR}{Text}

\item{design2x2}{Text}

\item{result}{Text}
}
\value{
Text
}
\description{
Omics-based interaction analysis by OBIF
}
