% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samHC.R
\name{samHC}
\alias{samHC}
\title{Sample Hierarchical Clustering}
\usage{
samHC(dataset, col.samples)
}
\arguments{
\item{dataset}{Text}

\item{col.samples}{Text}
}
\value{
Text
}
\description{
Sample Hierarchical Clustering
}
