% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exGSE.R
\name{exGSE}
\alias{exGSE}
\title{Export to GSEA}
\usage{
exGSE(final.labels, col.ft.name, dataAR)
}
\arguments{
\item{final.labels}{Text}

\item{col.ft.name}{Text}

\item{dataAR}{Text}
}
\value{
Text
}
\description{
Export to GSEA
}
