% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fDisc.R
\name{fDisc}
\alias{fDisc}
\title{Feature Discovery by OBIF}
\usage{
fDisc(final.results, l2fc.co, pval.co, pval.sel)
}
\arguments{
\item{final.results}{Text}

\item{l2fc.co}{Text}

\item{pval.co}{Text}

\item{pval.sel}{Text}
}
\value{
Text
}
\description{
Feature Discovery by OBIF
}
