% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tresh.R
\name{tresh}
\alias{tresh}
\title{RPKM/FPKM/TPM Tresholding}
\usage{
tresh(dataAR, treshold)
}
\arguments{
\item{dataAR}{A numeric data matrix with the analyis-ready feature expression values where row are features and columns are samples.}

\item{treshold}{A numeric value with the desired treshold to apply to the dataset.}
}
\value{
A data frame with the individual p-values per feature for main effects, simple main effects and interaction effects.
}
\description{
The tresh function is used to curate the dataset to include only the features that meet the treshold criteria across all samples.
The treshold is user-defined based and relative to the expression values in the dataset as PRKM/FPKM/TPM.
This function does not perform transformations on the expresion values.
}
